Refactor the project to use node/webpack/makefile
There was already a need to import multiple js files in html example2d,
and that was going to go crazy soon ! So now, with webpack, it's all
good.

I also added a makefile for basic commands (test and build) because
makefile is more common than npm or anything and could be hmogeneous
with other projects on other stacks.